refactor: 💡 extract to helper function
refactor: 💡 rename for readability
